changed short to int as in include.h

+++ plat.c
@@ -24,7 +24,7 @@ short vlist[MVERT], clist[MCELL], blist[MBORD], bublist[MCELL], vnbr[MVERT][3],
       vper[MVERT][3], cadj[MVERT][3];
 short nbsides[MBORD], ncsides[MCELL], iel[MELOST];
 short nv, nc, nb, nbub=0, onv, onc, onb, nel;
-short elosscount, bpinchcount;
+int elosscount, bpinchcount;
 REAL tfoam=0.0, boxwid=1.0, boxhgt=1.0, netenergy, henckyeps, volfrac=0.0;
 REAL lengthdelta, pressuredelta, equilsup, filmwid, minvvlen,
       diffuserate, bprelax, areasup, bareasup, minbfrac, maxbfrac, maxdv,
